used html5-camera instead of react-webcam
Now, the button is circular and is animated on click. The image currently takes up the entire page, but I want to play with the dimensions once we can test it on mobile, because that's how most people will be using the app. Also had to rename the Camera component to be "Webcam" so i changed that in the home.jsx